client_api/requirements.txt
bcrypt==3.2.0
CurrencyConverter==0.14.3
facepy==1.0.12
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Blogging==1.2.2
Flask-Caching==1.9.0
Flask-FileUpload==0.5.0
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
github-api-v4==0.0.0
google-api-core==1.23.0
google-auth==1.22.1
google-cloud==0.34.0
google-cloud-core==1.4.3
google-cloud-datastore==1.15.3
google-cloud-ndb==1.7.0
google-cloud-pubsub==2.1.0
google-cloud-tasks==2.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.33.1
Jinja2==2.11.2
Markdown==3.3.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.13.0
PyGithub==1.54.1
pygorithm==1.0.4
PyJWT==1.7.1
PyMySQL==1.0.2
redis==3.5.3
requests==2.24.0
requests-toolbelt==0.9.1
shortuuid==1.0.1
slack-sdk==3.4.0
SQLAlchemy==1.3.23
uuid==1.30
Werkzeug==1.0.1
WTForms==2.3.3
python-decouple
